perform semiempirical calculations on multiple lowest energy confs from FF

+++ lib/semiempirical_opt.py
@@ -7,40 +7,52 @@ import traceback
 import numpy as np
 
 from .log_parser import XtbLog, G16Log
-from .file_parser import mol2xyz, xyz2com, write_mol_to_sdf
+from .file_parser import mol2xyz, xyz2com, write_mol_to_sdf, write_mols_to_sdf
 
 
-def xtb_optimization(folder, sdf, xtb_path, rdmc_path, g16_path, level_of_theory, n_procs, job_ram, base_charge, mult):
+def semiempirical_optimization(folder, sdf, xtb_path, rdmc_path, g16_path, level_of_theory, n_procs, job_ram, base_charge, mult, method):
     basename = os.path.basename(sdf)
     file_name = os.path.splitext(basename)[0]
-    scratch_dir = os.path.join(folder, file_name)
 
     parent_dir = os.getcwd()
+    child_dir = os.path.abspath(os.path.join(folder, file_name))
+    os.chdir(child_dir)
 
-    os.chdir(scratch_dir)
+    if rdmc_path and method == "GFN2-XTB":
+        mols = Chem.SDMolSupplier(sdf, removeHs=False, sanitize=False)
+        os.remove(sdf)
+        conf_ids_ens = []
+        for conf_ind, mol in enumerate(mols):
+            scratch_dir = f"{file_name}_{conf_ind}"
+            os.makedirs(scratch_dir, exist_ok=True)
+            os.chdir(scratch_dir)
 
-    if rdmc_path:
-        mol = Chem.SDMolSupplier(sdf, removeHs=False, sanitize=False)[0]
-        xyz = mol2xyz(mol)
+            xyz = mol2xyz(mol)
 
-        g16_command = os.path.join(g16_path, 'g16')
-        head = '%nprocshared={}\n%mem={}mb\n{}\nexternal=\"{}/rdmc/external/xtb_tools/xtb_gaussian.pl --gfn 2 -P\"\n'.format(n_procs, job_ram, level_of_theory, rdmc_path)
+            g16_command = os.path.join(g16_path, 'g16')
+            head = '%nprocshared={}\n%mem={}mb\n{}\nexternal=\"{}/rdmc/external/xtb_tools/xtb_gaussian.pl --gfn 2 -P\"\n'.format(n_procs, job_ram, level_of_theory, rdmc_path)
 
-        comfile = file_name + '.gjf'
-        xyz2com(xyz, head=head, comfile=comfile, charge=base_charge, mult=mult, footer='\n')
+            comfile = f"{file_name}_{conf_ind}.gjf"
+            xyz2com(xyz, head=head, comfile=comfile, charge=base_charge, mult=mult, footer='\n')
 
-        logfile = file_name + '.log'
-        outfile = file_name + '.out'
+            logfile = f"{file_name}_{conf_ind}.log"
+            outfile = f"{file_name}_{conf_ind}.out"
 
-        with open(outfile, 'w') as out:
-            subprocess.run('{} < {} >> {}'.format(g16_command, comfile, logfile), shell=True, stdout=out, stderr=out)
-            os.remove(sdf)
+            with open(outfile, 'w') as out:
+                subprocess.run('{} < {} >> {}'.format(g16_command, comfile, logfile), shell=True, stdout=out, stderr=out)
 
-        log = G16Log(logfile)
-        conf = mol.GetConformer()
-        for i in range(mol.GetNumAtoms()):
-            conf.SetAtomPosition(i, log.Coords[i,:])
-        write_mol_to_sdf(mol, f'{file_name}_opt.sdf')
+            log = G16Log(logfile)
+            conf_ids_ens.append((conf_ind, log.E))
+            conf = mol.GetConformer()
+            for i in range(mol.GetNumAtoms()):
+                conf.SetAtomPosition(i, log.Coords[i,:])
+
+            os.chdir(child_dir)
+            shutil.rmtree(scratch_dir)
+
+        write_mols_to_sdf(mols, f'{file_name}_confs.sdf')
+        conf_ids_ens.sort(key=lambda x: x[1])
+        write_mol_to_sdf(mols[conf_ids_ens[0][0]], f'{file_name}_opt.sdf')
         
     else:
         xtb_command = os.path.join(xtb_path, 'xtb')
